requirements.txt
google-adk[a2a]
uvicorn[standard]
requests

